feat: UML을 기반으로 Operator, ExpressionParser, Formula 타입 생성

+++ Calculator/Calculator/CalculateItem.swift
@@ -8,7 +8,7 @@ extension Int: CalculateItem {
     
 }
 
-extension Float: CalculateItem {
+extension Double: CalculateItem {
     
 }
